SmartBody: Refactored BML code to use smart pointers
INTERIM BUILD!

Trying to prevent any more / new memory leaks as I reorganize the BML processor for interruption.  Most prominent use relates to BmlRequest, TriggerEvent, and SynchPoint.

This version still has lots of problems, but its able to make it through the AAMAS 2008 demo seq, so it is mostly usable in a development environment (hence the check-in).  The biggest problem is the circular references in the SynchPoint doubly-linked list (and hence a memory leak).  This will be replaced with a ordered list in each BehaviorReqest.

I need to write a doc & example on everything I learned about using smart pointers and the idiom I developed for my BmlRequest->TriggerEvent->SynchPoint "tree" nodes.  The use of weak_ptr for back references does solve a lot of problems I had with the srShared class for MeControllerContext objects.