SmartBody-lib: Removed #define SYNC_LINKED_LIST and all references
This removes code that ordered SynchPoints via a linked list, but also created a memory link when using smart pointers.

Only SpeechRequest properly replaced the link list with a std::vector for ordering, and then only for time markers.  Need to standardize this across all BehaviorRequest objects (replacing the implied ordering of the standard markers)