stroke-start and stroke-end are now stroke_start and stroke_end, per recent BML workshop meeting.
Experiment in not initializing SynchPoints with a name value.  Wrapped by SYNC_NAME_AT_INIT, currently set to 1 (original codebase).  This effort arises out of two reasons:
  A SynchPoint object can have multiple names, via references.
  BmlRequest::getSynchPoint(..) was creating and registering new SynchPoint objects, leaving side effects if the behavior failed to parse.

More to the point, this is also an effort to fix bug 1968966.  It is preventing me from build a proper regression test, and BmlRequest::getSynchPoint(..) sits at the heart of it.