Committing an implementation of the RemoteSpeech protocol for CereVoice speech synthesis.
This package does not include CereVoice.  It is a commercial product, and you will need to purchase a license to their SDK to use this code.  However, we expect it may be instructive to other developers who are interested in building their own RemoteSpeech services.

Please see the README.txt for more details.