speech_audiofile - complete re-write of this class.
This class functions the same EXCEPT:
It will now look for a .bml file the same base name as the audio file.  This file will contain the word break timing information and viseme information.  We currently have a tool (called word_breaker) that creates this file by taking the text and sending it through a TTS engine, getting a rough timing of the word break information.

This current implementation still relies on the .ltf file for the phoneme schedule, and converts them to visemes at runtime.