exporter - changes made by Kristin Parker, cica July 2008. These changes split up the skm exporter into 2 main files: one that contains all of the export logic and the other that contains the GUI code.
This gives the ability to export .skm files from the command line in a batch fashion.

The other files are support files for the batch export process.

There are 2 ways to export .skms through batch:  the default way that exports all .skms contained in the given maya file, or the single export way which only exports a single specified animation contained in the maya file.

The single method also allows you to specify a "base" joint for exporting partial bone animations (not fully tested, but should work)