exporters - changes made to the .sk exporter by Kristin Parker, circa July 2008. These changes are similar to the skm exporter in r189, splitting up the exporter into two parts, one that does the exporting, and the other handling displaying the GUI.  This is done so that you can export an .sk from the command line as a batch process.
These modifications also include saving the .sk metadata into custom data properties (similar to how it's done with the .skm information).  This is so that you can set the .sk data, and it's stored inside the maya file.

Other options include exporting with and without limit data, and also choosing whether or not to export the limits as set in the maya file, or setting them to 'free'.