word_breaker - new offline tool, by Arno Hartholt, for the speech_audiofile code path in sbm.
From the initial commit (only change is that files are stored as .xml):
Word Break Boundary Timing Acquirerer Module. Word Breaker for short. Takes as input pairs of prerecorded audio file names and utterance text, sends it to a TTS module, and parses the reply in order to create BML containing the timing of word boundaries. These are hopefully approximate to the timing in the prerecorded audio, and can be tweaked later. The BML is saved in a .wbb file that has the same name as the audio file. Currently, wbb files only contain word timing information. Viseme timing information  needs to come from .ltf files.

To do:
- provide correct values for stroke and relax attributes
- provide command line parameters