Added MeDefaultPrunePolicy.
This new prune policy propagates the policy of children to the parent.  If all a controller has a prune policy that prevent pruning, and all ancestor controllers use this new prune policy, then the ancestors will also discourage pruning.  Since this logic is in its own policy object, it is easy to replace this default behavior with a new prune policy.