Isolated examples of the blend curve errors.
test-blend-curve-ready-stroke-anim.xml syncs the ready and relax animation to two word breaks of a speech behavior.  The current results is blend curve that blends in, blends out, and blends back in.

The final blend-in point is usually scheduled at about double the time of other sync points, resulting in a very slow return blend in.  This is strongly similar to the behavior seen with event behaviors synced to speech:relax or speech:end (and often missing after cleanup during "vrAgentBML ... end").

If the controller in a motion controller with source animation that ends in the base posture, this may not be obvious, as the time curve locks the animation on the final frame.  However, the motion will remain active, and obscure some or all of the current posture.  Additionally, the motion will not be pruned because the prune algorithm will see the controller as active.

This behavior has been seen in similarly synchronized head nod controllers, but the test-blend-curve-ready-relax-head-nod.xml does not realize the full error.