Using SyncPoint parent references with zero offsets instead of shared SyncPoint objects.
This created a lot of new warnings of unset parents and the like.  To help me track them down, I've included the BehaviorRequest::unique_id in these warnings.  The results are very verbose, but bear with me while I clean it up.

Also, I see the code for relative timing should be active, but it doesn't seem to be parsing correctly.  I need to investigate that.