Fixed missing prune call, and thus the accumulation of controllers.
Moved the prune call was moved from the cleanup sequence, which was aborted before being run, to an execute_later call in the cleanup() function.

I also moved the command to abort the cleanup sequence from BmlRequest::cleanup() to ::unschedule(), so it will be called during an interrupt, removing the "vrAgentBML ... end complete" (already replaced with "vrAgentBML ... end interrupted").  Additionally, I removed the command to abort the start sequence from the cleanup command as extraneous.

I'm seeing vrSpoke more often, but not always.  Next time on the agenda.