branch mel script - optimization. $names = `keyframe -q -name $jointName`; This line was being called per joint, per frame.
This could be done once at the script start since this data is static, not frame specific.  The problem is that there isn't a 2D string array data type in MEL, and you can't create custom data types.

I've modified the script to pull this data at script start, and store it in a regular string array, and access it with [(i*width)+j], emulating a 2d array.

By adding this ugly hack, it has sped up the exporter dramatically.  A 150 frame anim that took 30 seconds on skm export now takes 4 seconds.