First pass at cleaning up the viseme code since merging the facebone code.
* Uses std::map:find(..) instead of iteration
* Removed confusing and buggy non_cap_viseme (we'll revisit case sensitivity when the bone bus visemes are standardized caps or lowercase)
* Fixed a couple of bugs in terms of memory leaks and forgotten variable renames of copy/pasted code.
* Renamed the maps to be more descriptive.

Tested with and without -facebone
When using facebone, remember to initialize the face animation data before creating the characters.  If using our face animations, use the "data/sbm-test/scripts/default-face-init.seq".  The line to call this in "data/sbm-test/scripts/default-init.seq" is currently commented out, because it interferes with eyeball rotation (our face animations accidentally overwrite the eyeball rotations; waiting on a fix to the exporter to filter these joints out).