tesAdded "test locomotion" rotational commands.
Fixed the interpretation of the rotation in MeCtLocomotionSimple.  It is applies rotation to the world_offset, instead of the translational offset.  This unnatural with constant, simultaneous translation and rotation, but is the nature of the inputs.  I'll look into implementing something like MeCtNavCirlce, with a command like "test locomotion circle <forward velocity> <angular velocity>".  This might be useful enough to apply to pawns as well.

Bunch of new commands or command variants:
  test locomotion  // prints syntax help with lots of chainable arguments
  set character <agent id> world_offset origin // translates to origin
  set character <agent id> world_offset zero  // zeros out world_offset transform
And new shorthand notations:
  test loco
  test loco char <agent id> dps <angular vel in degrees per sec>
  set char <agent id> w_o origin h 90// it didn't used to take w_o here, chained with heading

Added macros DEGF and RADF to gwiz, float variants of DEG and RAD.