Pluggable per-behavior scheduling modules.
The ugly scheduling logic is now wrapped into BehaviorScheduleLinear, and is replaceable with other types of BehaviorScheduler objects.

It is the responsibility of the BehaviorRequest factory function to select the BehaviorScheduler, but some top level BehaviorRequests, like MeControllerRequest and SequenceRequest, have defaults.

SpeechRequest implements it own scheduling, and is not using a BehaviorRequest (but I might wrap the existing scheduling later).

Limited testing, but I succeeded in getting through SASO (gaze, nod, speech, posture all seem to work fine).