sbm - breaking change in position data sent through bonebus. - bone positions are now sent across bonebus as x,y,z instead of some arbitrary notation used for legacy reasons.  Also, there is no longer a special case for the "base" bone. - world_offset positions are sent as x,y,z as well - CommAPI setcameraposition messages are now also sent as x,y,z.  Unneeded CommAPI messages have been removed.  The only CommAPI messages are now setcameraposition and setcamerarotation
This is a breaking change because of how you interpret your bonebus data on the renderer side.  You will need to adjust your handling of this data on the renderer side or else you will get strange results.  Also, any scripts that use the CommAPI setcameraposition function will need to be changed as well.

The next round of changes will involve the rotation function, with similar changes planned.