sbm - breaking change in rotation data sent through bonebus.  This is the similar change that was made to positions in r547 - bone rotations are now sent across bonebus as w,x,y,z (quat) instead of some arbitrary notation used for legacy reasons.  Also, there is no longer a special case for the "base" bone. - world_offset rotations are sent as w,x,y,z (quat) as well - CommAPI setcamerarotation messages are now sent as x,y,z (euler).
This is a breaking change because of how you interpret your bonebus data on the renderer side.  You will need to adjust your handling of this data on the renderer side or else you will get strange results.  Also, any scripts that use the CommAPI setcamerarotation function will need to be changed as well.