Replace blink viseme command with permanent blink controller.
This update reverts some of R578, so that SbmCharacter::eye_blink_update(..) again use the bonebusCharacter (if available).

The primary reason for this change was to prevent the influx of tiny individual controllers caused by blink.  Each blink was generating 12 controllers! (Three per command, to open and close, per eye.)  In part this is a general issue for the "char * viseme" command to be addressed later.

This replacement runs independent of the eye_blink_update(..).  The blink cycle is unsynchronized as a result.  However, I see no case where both would be visible (facebone occludes bonebus commands in those renderers that happen to implement both).

The new blink controller sits before the head controller group, such that manipulating AU_45 will not be interrupted by this controller.  (This raises issues with how to revert to blinking after such commands.)

I pulled the constants out of the old blink function and use them in both blink implementations.  I also updated them to be more realistic.  I have not verified the results on a bonebus based renderer.

The current periodic controller looks a little too regular.  I'm considering replacing this with an MeCtAperiodicReplay, matching the current bonebus blink algorithm.