Renamed SyncPoints to SequenceOfNamedSyncPoints.
(Verbose name, but I wanted to convey it is the SyncPoint that is named, not the Sequence itself.)

First of a two-part refactor to make the SyncPoint name/id better bound to the actual SyncPoint objects.

The end goal is to clean up the existing and upcoming scheduler code, which obviously uses the naming and ordering of SyncPoints heavily.