Changed facebone viseme blending to be a percentage of the viseme duration instead of a fixed number (0.1 seconds). Percentage is now around 140, in order to smooth out viseme intensive schedules. Might need some tweaking.
Background: SBM facebone currently expects visemes to be single frame animations, i.e.d not containing a curve with ready, stroke, relax, etc. Because of this, the regular blending algorithm isn't used here. Instead, visemes are linearly blended between weight 0 and 1 over a certain duration (and then back to 0), which now is a percentage of the duration of that viseme.