export_skm - added skmMergeAnimLayers() which will merge the animation layers before exporting.
This is the same as selecting all anim layers, then going to Layers->Merge Layers.

This is a required step if the animation has layers otherwise you won't get correct data.  Unfortunately, this modifies the Maya file.  So, now SKM exporting a file is a destructive step and the Maya file needs to be re-loaded, otherwise you will lose your layers.

If there is a way to get the correct data out of Maya when layers are present, then we will change this behavior.  But until then, the merge will take place.