Update the default attributes from character to controllers. Automated the whole process, so now one can use "addDefaultAttribute[Type]" in the controller constructor to specify the variable-attribute pair. This simplify the work to expose some controller parameters in the resource viewer for tweaking.
Currently, only exposed some parameters in saccade and reach controllers. But more can be added in the future iterations.

Change the MeController::init() to accept SbmPawn* as a parameter, and change all controller codes accordingly. These changes are necessary so it will be less awkward to connect the attributes between the character and the controller.