Initial update of the reach controller for zebra 2.0 skeleton. The pure IK reach seems to work fine, though more animation data is needed to test the example-based reach.
Fixed the problem of inconsistent pawn positions when reaching. This problem causes problematic reaching results in Unity ( and Ogre ? ). The culprit is that the gmat is not updated after set pawn world_offset. For some reason this problem never occurred to me when testing in SBM viewer. Guess it's because I always drag the pawn directly to set the position.