Update more changes in physics based character. Still in progress, so all related code are disabled.
Fixed(guess?) the crash issue when initializing the shadow map. I guess the problem happens when the GLSL shader is not supported by the graphics hardware. This would cause shader deletion to crash. If the crash still occurs after this fix, I may need to test on that specific machine to nail down the problem.