Sync points from audiofiles are now included in the BML request, so behaviors can be automatically synced with it. For example, you can specify:
<speech ref="utterance1" id="myutterance"/><head type="NOD" start="myutterance:T8"/>

and the head nod will be synced with the fourth word (each word has two marks, for the beginning and the end of the word). Previously, this was not possible. Note thet the 'id' tag in the reference <speech> tag will replace any "id" tag that exists in the generated BML.