Fixes for vs2008 by removed references to sbm-debugger.
Added attributes to pawn for handling collision shape for physics, reaching and steering:
"collisionShape" - can be sphere, box, capsule or null
"collisionShapeScale" - size of that object
"createPhysics" - creates a physics representation of that pawn
"enablePhysics" - enables or disables the physics-based representation of that pawn

Some refactoring of physics and reaching code. More needs to be done.