Fix the save function for SkMotion : also write out sync point when saving out a motion file.
Partially fix the IK for pre-rotation/orientation skeleton. Now include joint orientation when computing global position. This fixes most problems with Adapt character, but for some reason the IK result is still not accurate enough. Could be missing something in the IK computation to account for pre-rotation. More fixes need to be done to figure this out.