converted externals definition to new svn 1.6.x style
was holding off on this because the new style is ugly.  but TortoiseSVN 1.7.x clients use this natively.  So, it would happen sooner or later through someone's commit.

Using peg revision only.  Not operative revision.  This is confusing.  But hopefully less so if we always use peg revision.

Links:

http://tortoisesvn.net/docs/nightly/TortoiseSVN_en/tsvn-dug-externals.html
http://svnbook.red-bean.com/en/1.7/svn.advanced.pegrevs.html#idp8910176
https://groups.google.com/forum/?fromgroups=#!topic/tortoisesvn-dev/ludN2m-YHyY

"And when no operative revision is provided, it defaults to being the same revision as the peg revision."