Add initial implementation for post-IK processing for locomotion. Now will adjust character's root position and legs based on terrain heights. The current method is still crude and will result in artifacts when the slope is steep.
Add foot step prediction in motion analysis. This will infer the foot position for the next step. Could be used to better adjust the motion. Also add the code to visualize and debug the foot step positions during locomotion. Still in progress, so it is disabled for now.

Some minor fix in transforming the normal vectors. This will produce correct normals for some characters with root pre-rotation ( ex. RocketBox characters ).