Fixes problems for simulation consistency. Removed some implicit VHMsg calls from bml parsing so there is always a deterministic order of command execution. This solves a lot of inconsistent behavior when running the same simulation multiple times ( for example, Machinema editor ). Also adjust the steering logic for consistency and add a locomotionReset function to immediately terminate locomotion and its corresponding blends. Usage : bml char * <locomotion enable="reset"/>
Also add a time stepping function for SBSimulationManager. It allows step forward the simulation with a delta time, instead of requiring absolute time. This simplifies the logic to control the simulation with external app like Unity. Ideally, the Update function in SmartBody_dll should use time stepping instead of absolute time, but right now it is left unchanged for compatibility with other projects.