Fixed a bug with initializing vertex buffer with GPU deformable model. This causes potential crash when there are more than one OpenGL context. Add default names when creating pawns/characters. Remove create/terrain menu item since it's not actively used.
Add a temp solution for refreshing/reloading all tool windows when create a new scene. The temp solution is to delete all existing tool windows so they are recreated after the new scene. To-do : find out or add proper refreshing function for each window so they are consistent with any modification to the current scene. 

Change project setting in debug mode. For some reason the compiler is complaining about obj size, so need to add /bigobj flag in debug mode to compile.