Change GPU deformable model to allow up to 4 lights. Change the parameter "NumLight" in shader code to adjust the lighting setting. Note : some graphics hardware seems to have problems handling too many variables. So limit the setting to two lights by default.
Add a quick fix for hair rendering by disable depth testing for hair geometry with alpha < 1.0 This prevents strange white color due to incorrect depth ordering.

Add a bool attribute "isReaching" for SBCharcter. This attribute indicates whether a character is in reaching. Would be helpful for animation planning when grasping cloth. Still need testing to see if it works.