Add more exception handling for steer agent. Need more test to see if this improves the random crash.
Initial attempt to add a turning angle epsilon in steering agent to prevent small and high frequency turning. It doesn't seem to remove all the jittering. Need more testing to verify if the same epsilon threshold need to be considered for speed and scoot as well. ( Though it seems there is already a speed and scoot threshold handling this ).