Some more fixes in Ogre renderer to render skin weights correctly.
Add initial implementation for skinning weight computation. The new method produces better skinning with less body shrinking artifacts. However, the method right now is slower than original method since it needs to smooth the skinning weights for the full mesh. 

The future plan for improving performance is to change the method to work fully in voxel space so it is resolution independent. For now, it is added as an option for computing auto-rigging.