Fix a very weird bug, which will cause fail to apply skeleton for some model
when a model need set meshScale, it will cause a problem.
(since fuse model dont need to set scale, this bug didnt show up)

in updateSkin() in sb_deformable_mesh.cpp.

There is like vPos=restPos[i]*_meshScale[0];

I think we dont need this, because the meshScale already be multiplied in updateTransformBuffer() function

so, for web version, it actually  multiply meshScale twice.
(one is multiply with matrix(bindPoseMat), one is multiply with vector3(vPos), cause a very weird result)

In pc version, this have no impact on the result which is weird

I delete the *_meScale[0] work fine, but I still keep it for now in case other platform need this.